+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 ## [Unreleased]
+### Added
+- functions for dealing with loacl network interfaces
+
 ### Fixed
 - resource-as-file function works properly now